% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInput.R
\name{createInput}
\alias{createInput}
\title{Create Input}
\usage{
createInput(percentFemale = 0.5, timeHorizon = 20,
  costDiscount = 0.03, qalyDiscount = 0.03, baselineAge = 40)
}
\arguments{
\item{percentFemale}{decimal: between 0 and 1, percentage of population that is female}

\item{timeHorizon}{integer: time in years for model}

\item{costDiscount}{decimal: between 0 and 1, percentage which payoffs decrease every year}

\item{qalyDiscount}{decimal: between 0 and 1, percentage which QALYs decrease every year}

\item{baselineAge}{integer: mean age at baseline}
}
\description{
Allows users to customize input parameters
}
