% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_default_settings}
\alias{get_default_settings}
\title{Get Default Settings}
\usage{
get_default_settings()
}
\value{
default settings

\itemize{
\item record_mode:
\item events_to_record:
\item agent_creation_mode:
\item update_continuous_outcomes_mode:
\item runif_buffer_size:
\item rnorm_buffer_size:
\item rexp_buffer_size:
\item agent_stack_size:
\item event_stack_size:
}
}
\description{
Exports default settings for the dynamic microsimulation
}
