% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plotCOPD}
\alias{plotCOPD}
\title{Plot COPD Occurrence/Incidence}
\usage{
plotCOPD(nPatients = 10000, argX = "age", argY = "incidence",
  aggregateBy = "none", groupBy = NULL, perCapita = NULL)
}
\arguments{
\item{nPatients}{number of patients/agents/individuals to simulate}

\item{argX}{string: one of "year", "age"}

\item{argY}{string: one of "incidence", "prevalence"}

\item{aggregateBy}{string: one of "age", "none"}

\item{groupBy}{string: one of "sex", "age", "gold"}

\item{perCapita}{numeric: if null, parameter is ignored, otherwise, population size to look at}
}
\value{
plot
}
\description{
Creates a plot of exacerbations
}
