% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plotExacerbations}
\alias{plotExacerbations}
\title{Plot Exacerbations}
\usage{
plotExacerbations(nPatients = 10000, argX = "year",
  argY = "exacerbation_rate", groupBy = "sex", exacType = c("all",
  "severe"), perCapita = NULL)
}
\arguments{
\item{nPatients}{number of patients/agents/individuals to simulate}

\item{argX}{string: one of "year"}

\item{argY}{string: one of "exacerbation_rate", "number_of_exacerbations"}

\item{groupBy}{string: one of "sex", "age", "gold"}

\item{exacType}{string: "all" or "severe"}

\item{perCapita}{numeric: if null, parameter is ignored, otherwise, population size to look at}
}
\value{
Returns a plot
}
\description{
Creates plots of exacerbations
}
